GameStore.WUI/Models/Concrete/Cart.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;

namespace GameStore.WUI.Models.Concrete
{
    public class Cart
    {
        private List<CartLine> lineCollection = new List<CartLine>();

        public void AddItem(Game game, int quantity)
        {
            CartLine line = lineCollection
                .Where(g => g.Game.Id == game.Id)
                .FirstOrDefault();

            if (line == null)
            {
                lineCollection.Add(
                    new CartLine
                {
                    Game = game,
                    Quantity = quantity
                });
            }
            else
            {
                line.Quantity += quantity;
            }
        }

        public void RemoveLine(Game game)
        {
            lineCollection.RemoveAll(l => l.Game.Id == game.Id);
        }

        public double ComputeTotalValue()
        {
            return lineCollection.Sum(e => e.Game.Price * e.Quantity);

        }
        public void Clear()
        {
            lineCollection.Clear();
        }

        public IEnumerable<CartLine> Lines
        {
            get { return lineCollection; }
        }
    }

}